refactor(review): split review_phase.py into package for file-size discipline (T36, advisor-zpv)
The original ``src/review_phase.py`` (3702 lines) is now a package:

* ``review_phase/_common.py`` (205 lines) — module-level constants
  (``_AdvisorRole``, ``_NON_VERDICT_SUMMARY_MARKERS``,
  ``_SELF_MOD_SYNTHESIS_PATTERNS``), OTel metric instruments
  (``_veto_retries_total`` / ``_veto_recovered_total`` /
  ``_veto_exhausted_total``), dataclasses (``ReviewGuardContext``,
  ``PreReviewContext``), and standalone helpers
  (``_is_meaningful_verdict``, ``_run_fallback_ingest_review``,
  ``_emit_advisor_loop_metric``, ``_detect_self_modification_context``).

* ``review_phase/_phase.py`` (3568 lines) — the ``ReviewPhase`` class
  body, structurally unchanged.

* ``review_phase/__init__.py`` (101 lines) — back-compat re-exports
  for the full public surface plus a ``__getattr__`` delegator to
  ``_phase`` for any patch site we didn't enumerate.

This is the minimum-viable split (Step 5 of the M2 plan) — a full
mixin-based decomposition isn't realistic given how heavily the
``ReviewPhase`` methods cross-reference per-PR state set in
``__init__`` (``self._advisor_attempt``, ``self._advisor_results``,
``self._advisor_pre_flight_plan``, ``self._post_verify_runner``,
``self._reviewers``, etc.). The shape we shipped extracts what *can*
be cleanly separated — module-level constants, regex patterns, metric
instruments, dataclasses, and stateless helpers — without breaking
any caller. A future ADR can revisit a deeper mixin split if the
``_phase.py`` file size becomes a problem.

Back-compat is total: every previously-importable symbol still
imports from ``review_phase`` without changes.

  from review_phase import ReviewPhase                  # works
  from review_phase import PreReviewContext             # works
  from review_phase import _is_meaningful_verdict       # works
  from review_phase import _emit_advisor_loop_metric    # works
  from review_phase import _veto_retries_total, ...     # works

Test patches against ``review_phase.analyze_patterns`` and
``review_phase.record_harness_failure`` continue to work because
``_phase.py`` looks up those two symbols via ``review_phase.<name>``
at call time (late-binding through the package namespace, which is
what ``unittest.mock.patch`` mutates). The pattern is documented at
the ``import review_phase`` line in ``_phase.py``.

Two regression-test path updates were required:

* ``tests/test_docstrings_issue_2405.py`` — reads
  ``review_phase/_phase.py`` instead of the missing ``review_phase.py``.
* ``tests/regressions/test_issue_6752.py`` — anchors point at
  ``review_phase/_phase.py``; xfail strict=False so the now-stale line
  numbers stay tolerated.

Refreshed ``docs/arch/generated/`` (arch-regen picked up the new
package boundary).

Verification:
- All 290+ review_phase / advisor / wiki / visual / hooks tests pass.
- ``make quality``: 12476 passing; only ``test_planner.py::test_build_prompt_truncates_long_body``
  fails (pre-existing flake, unchanged by this refactor).
- ``ruff check`` and ``pyright`` clean on ``src/review_phase/``.
