+++ docs/arch/.meta.json
@@ -1,33 +1,33 @@
 {
-  "regenerated_at": "2026-05-10T22:21:29.179271+00:00",
-  "commit_sha": "109888f05b610a37b5f64125c97c56f08c8a23ec",
+  "regenerated_at": "2026-05-11T15:51:04.049330+00:00",
+  "commit_sha": "d649803df37eb1b898b026768dbf91109be77938",
   "artifacts": {
     "loops.md": {
-      "source_sha": "109888f05b610a37b5f64125c97c56f08c8a23ec"
+      "source_sha": "d649803df37eb1b898b026768dbf91109be77938"
     },
     "ports.md": {
-      "source_sha": "109888f05b610a37b5f64125c97c56f08c8a23ec"
+      "source_sha": "d649803df37eb1b898b026768dbf91109be77938"
     },
     "labels.md": {
-      "source_sha": "109888f05b610a37b5f64125c97c56f08c8a23ec"
+      "source_sha": "d649803df37eb1b898b026768dbf91109be77938"
     },
     "modules.md": {
-      "source_sha": "109888f05b610a37b5f64125c97c56f08c8a23ec"
+      "source_sha": "d649803df37eb1b898b026768dbf91109be77938"
     },
     "events.md": {
-      "source_sha": "109888f05b610a37b5f64125c97c56f08c8a23ec"
+      "source_sha": "d649803df37eb1b898b026768dbf91109be77938"
     },
     "adr_xref.md": {
-      "source_sha": "109888f05b610a37b5f64125c97c56f08c8a23ec"
+      "source_sha": "d649803df37eb1b898b026768dbf91109be77938"
     },
     "mockworld.md": {
-      "source_sha": "109888f05b610a37b5f64125c97c56f08c8a23ec"
+      "source_sha": "d649803df37eb1b898b026768dbf91109be77938"
     },
     "changelog.md": {
-      "source_sha": "109888f05b610a37b5f64125c97c56f08c8a23ec"
+      "source_sha": "d649803df37eb1b898b026768dbf91109be77938"
     },
     "functional_areas.md": {
-      "source_sha": "109888f05b610a37b5f64125c97c56f08c8a23ec"
+      "source_sha": "d649803df37eb1b898b026768dbf91109be77938"
     }
   }
 }
